refactor: TIAGo trajectory optimization: modularize classes and improve structure
- Removed unused imports and consolidated necessary ones.
- Transferred configuration and trajectory optimization classes to a base module.
- Enhanced the `OptimalTrajectoryIPOPT` class to extend from `BaseOptimalTrajectory`.
- Streamlined initialization and component setup for better clarity.
- Updated `TiagoTrajectoryIPOPTProblem` to inherit from `BaseTrajectoryIPOPTProblem` with specific configurations.
- Improved error handling and logging throughout the trajectory generation process.
- Simplified the objective and constraint functions for better performance.
